requirements.txt
absl-py==1.0.0
APScheduler==3.8.1
asgiref==3.4.1
astunparse==1.6.3
backports.entry-points-selectable==1.1.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.3
Django==3.2.7
django-anymail==8.4
django-cors-headers==3.10.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
filelock==3.1.0
flatbuffers==2.0
gast==0.4.0
grpcio==1.42.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
itypes==1.2.0
Jinja2==3.0.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
platformdirs==2.4.0
protobuf==3.19.1
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.2.0
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
stripe==2.63.0
termcolor==1.1.0
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.3
zipp==3.6.0
